Limpiando el contenido del directorio principal
