+++ requirements.txt
@@ -1,0 +1,5 @@
+asyncio
+jupyter
+ray[serve]
+requests
+StringGenerator
